+++ src/app/app-routing.module.ts
@@ -6,6 +6,7 @@ import {JobsListComponent} from './components/jobs-list/jobs-list.component';
 import {SettingsComponent} from './components/settings/settings.component';
 import {LogsComponent} from './components/logs/logs.component';
 import {S3ExplorerComponent} from './components/s3-explorer/s3-explorer.component';
+import {S3StatsComponent} from './components/s3-stats/s3-stats.component';
 
 const routes: Routes = [
   {
@@ -27,6 +28,10 @@ const routes: Routes = [
   {
     path: 's3-explorer',
     component: S3ExplorerComponent
+  },
+  {
+    path: 's3-stats',
+    component: S3StatsComponent
   },
   {
     path: 'logs',